requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
pytest==8.3.2
streamlit==1.38.0
pytest>=8.0